Conditionally display log components
