+++ app/src/js/utils/log.js
@@ -3,3 +3,5 @@
 export default function log () {
   console.log.apply(console, arguments);
 }
+
+export const metricsNotConfiguredMessage = 'Metrics not configured';
